Start the work of updating the analysis interface
This commit starts the work of implementing a hierarchy of analyses, including multi-event analyses.
The new concept(!) of an Analysis aims to generalise the concept of a "Production" from older versions of the code.
Hopefully this will also have the effect of removing some of the O3 PE-rota-centric language from the package.